refactor(web): move the branch transport below the hook, breaking the import cycle
`repo-coverage` refused the seam: BranchesBackendContext imports the backend,
the backend imported `hooks/useBranches.ts` for its request helpers, and that
module imports the context. A three-module value-import cycle.

I called this "backwards layering" in the commit that introduced it and filed
it as a follow-up. That was wrong twice over — it is a cycle rather than a
matter of taste, and it was mine to fix in the diff that created it. The
guard is what said so.

`buildBranchUrls`, `parseBranchesResponse`, `branchesApi` and the structured
`BranchApiError` move to `lib/branches-backend.ts`, beside the backend that
is now their only production caller. `useBranches.ts` keeps the React half
and re-exports the three types its consumers name, so no component changed.

keeper-parity moved with them, in both directions: `branches-backend.ts`
now reaches the daemon and is classified, `useBranches.ts` has stopped
reaching and its entry is gone. The ledger fails on either half, which is
how a stale entry gets noticed rather than inherited.

Also worth stating: the local pass that missed this ran apps/web's own
suites and not `arch-lint`, whose scans are repo-wide — so a new module in
apps/web is inside its area even though its files are not.
